refactor: split EpicGames facade methods into domain modules
Extract ~100 facade methods from lib.rs (1417 lines) into 7 domain
modules under src/facade/: auth, assets, account, fab, store, cosmos,
and misc. lib.rs now holds only the struct definition, constructor,
and tests (~175 lines). Mirrors the existing api/ split pattern.

Zero public API changes — all methods remain on EpicGames.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-opencode)
