+++ src/facade/mod.rs
@@ -1,0 +1,14 @@
+/// Authentication facade methods
+mod auth;
+/// Asset and manifest facade methods
+mod assets;
+/// Account, social, and entitlement facade methods
+mod account;
+/// Fab marketplace facade methods
+mod fab;
+/// Store, catalog, and commerce facade methods
+mod store;
+/// Cosmos / Unreal Engine facade methods
+mod cosmos;
+/// Miscellaneous facade methods (presence, cloud saves, service status)
+mod misc;
